feat: 세  저    스테  스트  ㄹ , 패ㅌ리  ㅇ

+++ TC275/infotainment_System.h
@@ -33,7 +33,7 @@ static char buf[32];
 
 void initInfotainment(void);
 void updateStateByPacket(const struct ActuatorPacket *packet);
-void updateStateByButton(unsigned int buttonState);
+void updateStateByButton(uint8 buttonState);
 void setActuatorPacket(struct ActuatorPacket* packet);
 
 void printInfoDisplay();
